fix(api): DeleteOrgDepartment TOCTOU guard + org departments .int() validation (#589, #592)

+++ api/src/interface/org/departments/[code]/route.ts
@@ -58,7 +58,7 @@ export const PUT = factory.createHandlers(
     z.object({
       parent_code: codeSchema.nullable().optional(),
       manager_employee_code: codeSchema.nullable().optional(),
-      order: z.number(),
+      order: z.number().int(),
     }),
   ),
   async (c) => {
